# docs/swagger.yaml
definitions:
...
  controllers.CreateDestinationAssetsInput:
    properties:
      destinationID:
        type: integer
      images:
        items:
          type: string
        type: array
      video_contents:
        items:
          $ref: '#/definitions/request.VideoInput'
        type: array
    type: object
...
  models.City:
    properties:
      id:
        type: integer
      lat:
        type: string
      long:
        type: string
      name:
        type: string
    type: object
  models.Destination:
    properties:
      address:
        type: string
      category:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      facilities:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      name:
        type: string
      operational_hours:
        type: string
      position:
        type: number
      ticket_price:
        type: number
      video_contents:
        items:
          $ref: '#/definitions/models.VideoContent'
        type: array
    type: object
  models.Image:
    properties:
      destination_id:
        type: integer
      id:
        type: integer
      url:
        type: string
    type: object
...
  models.VideoContent:
    properties:
      description:
        type: string
      destination_id:
        type: integer
      id:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  request.CreateDestinationInput:
    properties:
      address:
        type: string
      category:
        type: string
      city:
        type: string
      description:
        type: string
      facilities:
        type: string
      image:
        items:
          type: string
        type: array
      name:
        type: string
      operational_hours:
        type: string
      position:
        type: number
      ticket_price:
        type: number
      video_contents:
        items:
          $ref: '#/definitions/request.VideoInput'
        type: array
    type: object
...
  request.VideoInput:
    properties:
      description:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
host: www.tripwise.my.id
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Documentation TripWise Capstone Project - Team 3
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TRIPWISE API
  version: "1.0"
paths:
...
  /destination:
    get:
      consumes:
      - application/json
      description: Fetch the destinations between two cities, based on the origin
        and destination city names
      parameters:
      - description: Origin city name
        in: query
        name: origin
        required: true
        type: string
      - description: Destination city name
        in: query
        name: destination
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: City not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch destinations
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get destinations for a specific route
      tags:
      - Routes
    post:
      consumes:
      - application/json
      description: Create a new destination and associate it with a city, images,
        and video contents
      parameters:
      - description: Destination Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.CreateDestinationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Destination'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new destination
      tags:
      - Destinations
  /destinations:
    get:
      consumes:
      - application/json
      description: Retrieve destinations filtered by city and/or category
      parameters:
      - description: City to filter destinations by
        in: query
        name: city
        type: string
      - description: Categories to filter destinations by
        in: query
        name: category
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Filter destinations based on city and category
      tags:
      - Destinations
  /destinations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a destination and its associated images and video contents
      parameters:
      - description: Destination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a destination
      tags:
      - Destinations
    get:
      consumes:
      - application/json
      description: Fetch detailed information of a destination including city, images,
        and video contents
      parameters:
      - description: Destination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get destination details
      tags:
      - Destinations
    put:
      consumes:
      - application/json
      description: Update destination details including city, images, and video contents
      parameters:
      - description: Destination ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Destination Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.CreateDestinationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a destination
      tags:
      - Destinations
  /destinations/assets:
    post:
      consumes:
      - application/json
      description: Upload images and videos to a destination
      parameters:
      - description: Create Destination Assets
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateDestinationAssetsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create assets (images and videos) for a destination
      tags:
      - Destinations
  /destinations/most-viewed-videos:
    get:
      consumes:
      - application/json
      description: Fetch destinations ranked by the number of video views
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get most viewed video content
      tags:
      - Destinations
  /destinations/personalized:
    get:
      consumes:
      - application/json
      description: Fetch destinations based on the user's preferences and categories
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get personalized destinations for a user
      tags:
      - Destinations
...
  /videos:
    get:
      consumes:
      - application/json
      description: Fetch all video contents stored in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all video contents
      tags:
      - Video
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"